Now using a second thread to get data
